requirements.txt
aiohttp~=3.8.6                      # requests
aiosqlite == 0.18.0                 # database
colorama~=0.4.6                     # logging colors
heckbot
hikari-lightbulb == 2.3.1           # command handler
hikari-miru == 2.0.2                # modals and buttons
hikari[speedups] == 2.0.0.dev115    # discord framework
loguru == 0.6.0                     # logging
pydantic[dotenv] == 1.10.4
pynacl~=1.5.0
pynamodb~=5.5.0
pytest~=7.4.3
pytest-asyncio~=0.21.1
python-dotenv~=1.0.0
PyYAML~=6.0.1
setuptools==68.2.2
table2ascii~=1.1.3
types-PyYAML==6.0.12.12
